requirements.txt
Flask==3.1.1
tensorflow==2.18.0
numpy
pillow
python-dotenv
Flask-CORS